fix(Golpe): alterar o tipo da variável fatorDano de int para double, bem como o retorno do método getFatorDano.

+++ saintSeiya/Golpe.java
@@ -1,25 +1,27 @@
-public class Golpe {
+public class Golpe
+{
     private String nome;
-    private int fatorDano;
-    
-    public Golpe(String nome, int fatorDano) {
+    private double fatorDano;
+
+    public Golpe(String nome, double fatorDano)
+    {
         this.nome = nome;
         this.fatorDano = fatorDano;
     }
-    
-    public String getNome() {
+
+    public String getNome()
+    {
         return this.nome;
     }
-    
-    public int getFatorDano() {
+
+    public double getFatorDano()
+    {
         return this.fatorDano;
     }
-    
-    public boolean equals(Object object) {
+
+    public boolean equals(Object object)
+    {
         Golpe outroGolpe = (Golpe)object;
-        
-        return
-            this.nome.equals(outroGolpe.getNome())
-            && this.fatorDano == outroGolpe.getFatorDano();
+        return this.nome.equals(outroGolpe.getNome()) && this.fatorDano == outroGolpe.getFatorDano();
     }
 }
